ci: use actions/checkout@v4
now that the rss-to-issues action uses node 20, it's time to update the
actions we use for our CI jobs, to also use node 20.
actions/checkout@v2 in verify-dist.yml was so old that GitHub was forcing
it to run under node 16 insted of node 12 which it was designed to run
under.

actions/github-script does not support node 20 yet, though. [1]

[1] https://github.com/actions/github-script/issues/421

Signed-off-by: Matthias Aßhauer <mha1993@live.de>